percolation that compiles but don't infinite loop

+++ percolation/_tree.pxd
@@ -2,9 +2,9 @@ import numpy as np
 cimport numpy as np
 
 ctypedef packed struct HIERARCHY_t : 
-    int left_node
-    int right_node
-    double value 
+    long left_node
+    long right_node
+    float value 
     int cluster_size
