+++ percolation/_tree.pyx
@@ -2,8 +2,8 @@ import numpy as np
 cimport numpy as np
 
 HIERARCHY_dtype = np.dtype([
-    ("left_node", np.intp),
-    ("right_node", np.intp),
-    ("value", np.float64),
-    ("cluster_size", np.intp)
+    ("left_node", np.int_),
+    ("right_node", np.int_),
+    ("value", np.single),
+    ("cluster_size", np.intc)
 ])
